extract resource link path logic into hook
