+++ site/src/components/JSONFileLinks/index.tsx
@@ -1,22 +1,11 @@
 import CodeBlock from "@theme/CodeBlock";
-import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import styles from "./styles.module.css";
 import { ExternalLinkButton } from "../ExternalLinkButton";
-import { ApiResource, Manifest } from "../../../../types/api.types";
-import manifestJson from "../../generated/manifest.json";
-
-const manifest: Manifest = manifestJson;
+import { ApiResource } from "../../../../types/api.types";
+import { useResourcePaths } from "@site/src/lib/resource-paths";
 
 export const JSONFileLinks = ({ resource }: { resource: ApiResource }) => {
-  const docCtx = useDocusaurusContext();
-  const files = manifest[resource];
-  if (!files) {
-    throw new Error(
-      `Missing manifest entry for resource: ${resource}, cannot render JSONFileLinks`
-    );
-  }
-
-  const { unminified, minified, csv } = files;
+  const { unminified, minified, csv, siteUrl } = useResourcePaths(resource);
 
   return (
     <div>
@@ -40,13 +29,13 @@ export const JSONFileLinks = ({ resource }: { resource: ApiResource }) => {
         {minified && (
           <>
             <h3>Minified </h3>
-            <CodeBlock>{`${docCtx.siteConfig.url}/${minified.apiPath}/${minified.name}`}</CodeBlock>
+            <CodeBlock>{`${siteUrl}/${minified.apiPath}/${minified.name}`}</CodeBlock>
           </>
         )}
         {unminified && (
           <>
             <h3>Unminified </h3>
-            <CodeBlock>{`${docCtx.siteConfig.url}/${unminified.apiPath}/${unminified.name}`}</CodeBlock>
+            <CodeBlock>{`${siteUrl}/${unminified.apiPath}/${unminified.name}`}</CodeBlock>
           </>
         )}
       </div>
